+++ engine/shaderc/src/shaderc.h
@@ -29,9 +29,10 @@ namespace dmShaderc
 
     enum ShaderLanguage
     {
-        SHADER_LANGUAGE_NONE = 0,
-        SHADER_LANGUAGE_GLSL = 1,
-        SHADER_LANGUAGE_HLSL = 2,
+        SHADER_LANGUAGE_NONE  = 0,
+        SHADER_LANGUAGE_GLSL  = 1,
+        SHADER_LANGUAGE_HLSL  = 2,
+        SHADER_LANGUAGE_SPIRV = 3,
     };
 
     enum ShaderStage
