+++ engine/shaderc/src/shaderc_private.h
@@ -1,0 +1,67 @@
+// Copyright 2020-2024 The Defold Foundation
+// Copyright 2014-2020 King
+// Copyright 2009-2014 Ragnar Svensson, Christian Murray
+// Licensed under the Defold License version 1.0 (the "License"); you may not use
+// this file except in compliance with the License.
+//
+// You may obtain a copy of the License, together with FAQs at
+// https://www.defold.com/license
+//
+// Unless required by applicable law or agreed to in writing, software distributed
+// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+// CONDITIONS OF ANY KIND, either express or implied. See the License for the
+// specific language governing permissions and limitations under the License.
+
+#ifndef SHADERC_PRIVATE_H
+#define SHADERC_PRIVATE_H
+
+#include "shaderc.h"
+
+#include <spirv/spirv_cross_c.h>
+
+namespace dmShaderc
+{
+    struct ShaderContext
+    {
+        spvc_context   m_SPVCContext;
+        spvc_parsed_ir m_ParsedIR;
+
+        spvc_compiler  m_CompilerNone;
+        spvc_resources m_Resources;
+
+        ShaderReflection m_Reflection;
+    };
+
+    struct ShaderCompiler
+    {
+        ShaderLanguage m_Language;
+    };
+
+    struct ShaderCompilerSPVC
+    {
+        ShaderCompiler m_BaseCompiler;
+        spvc_compiler  m_SPVCCompiler;
+    };
+
+    struct ShaderCompilerSPIRV
+    {
+        ShaderCompiler m_BaseCompiler;
+    };
+
+    ShaderCompilerSPIRV* NewShaderCompilerSPIRV(HShaderContext context);
+    void                 DeleteShaderCompilerSPIRV(ShaderCompilerSPIRV* compiler);
+    void                 SetResourceLocationSPIRV(HShaderContext context, ShaderCompilerSPIRV* compiler, uint64_t name_hash, uint8_t location);
+    void                 SetResourceBindingSPIRV(HShaderContext context, ShaderCompilerSPIRV* compiler, uint64_t name_hash, uint8_t binding);
+    void                 SetResourceSetSPIRV(HShaderContext context, ShaderCompilerSPIRV* compiler, uint64_t name_hash, uint8_t set);
+    const char*          CompileSPIRV(HShaderContext context, ShaderCompilerSPIRV* compiler, const ShaderCompilerOptions& options);
+
+    ShaderCompilerSPVC*  NewShaderCompilerSPVC(HShaderContext context, ShaderLanguage language);
+    void                 DeleteShaderCompilerSPVC(ShaderCompilerSPVC* compiler);
+    void                 SetResourceLocationSPVC(HShaderContext context, ShaderCompilerSPVC* compiler, uint64_t name_hash, uint8_t location);
+    void                 SetResourceBindingSPVC(HShaderContext context, ShaderCompilerSPVC* compiler, uint64_t name_hash, uint8_t binding);
+    void                 SetResourceSetSPVC(HShaderContext context, ShaderCompilerSPVC* compiler, uint64_t name_hash, uint8_t set);
+    const char*          CompileSPVC(HShaderContext context, ShaderCompilerSPVC* compiler, const ShaderCompilerOptions& options);
+
+}
+
+#endif
